Refactoring of settings structure.
