+++ src/Lykke.Service.Assets/Settings/CqrsSettings.cs
@@ -1,6 +1,5 @@
 ﻿using Lykke.Common.Chaos;
 using Lykke.SettingsReader.Attributes;
-using System;
 
 namespace Lykke.Service.Assets.Settings
 {
@@ -10,7 +9,5 @@ namespace Lykke.Service.Assets.Settings
         public string RabbitConnectionString { get; set; }
         [Optional]
         public ChaosSettings ChaosKitty { get; set; }
-        public string QueuePostfix { get; set; }
-        public TimeSpan RetryDelay { get; set; }
     }
 }
